requirements.txt
django==1.11.18
MySQL-python==1.2.5
python-memcached==1.59
python-ldap==3.1.0
django-bootstrap-form==3.4
django-countries==5.3.2
django-recaptcha==2.0.2
django-imagekit==4.0.2
Pillow==5.4.1
django-taggit==0.23.0
python-magic==0.4.15
django-localflavor==2.1
django-formtools==2.1
